moves searchVector into 'search' function
