+++ packages/orama/src/methods/search.ts
@@ -9,7 +9,7 @@ import {
   getInternalDocumentId
 } from '../components/internal-document-id-store.js'
 import { createError } from '../errors.js'
-import { getNanosecondsTime, getNested, sortTokenScorePredicate, safeArrayPush } from '../utils.js'
+import { getNanosecondsTime, getNested, sortTokenScorePredicate, safeArrayPush, formatNanoseconds } from '../utils.js'
 import type {
   AnyOrama,
   BM25Params,
@@ -30,7 +30,7 @@ import type {
   TypedDocument
 } from '../types.js'
 import { MODE_FULLTEXT_SEARCH, MODE_VECTOR_SEARCH } from '../constants.js'
-import { searchVector } from './search-vector.js'
+import { findSimilarVectors } from '../components/cosine-similarity.js'
 
 const defaultBM25Params: BM25Params = {
   k: 1.2,
@@ -119,6 +119,63 @@ export async function search<T extends AnyOrama, ResultDocument = TypedDocument<
 
   throw ('No other search modes are supported yet')
 }
+
+export async function searchVector<T extends AnyOrama, ResultDocument = TypedDocument<T>>(orama: T, params: SearchParamsVector<T, ResultDocument>): Promise<Results<ResultDocument>> {
+  const timeStart = await getNanosecondsTime()
+  let { vector } = params
+  const { property, limit = 10, offset = 0, includeVectors = false } = params
+  const vectorIndex = orama.data.index.vectorIndexes[property]
+  const vectorSize = vectorIndex.size
+  const vectors = vectorIndex.vectors
+
+  if (vector.length !== vectorSize) {
+    throw createError('INVALID_INPUT_VECTOR', property, vectorSize, vector.length)
+  }
+
+  if (!(vector instanceof Float32Array)) {
+    vector = new Float32Array(vector)
+  }
+
+  const results = findSimilarVectors(vector, vectors, vectorSize, params.similarity)
+
+  const docs: Result<ResultDocument>[] = Array.from({ length: limit })
+
+  for (let i = 0; i < limit; i++) {
+    const result = results[i + offset]
+    if (!result) {
+      break
+    }
+
+    const originalID = getInternalDocumentId(orama.internalDocumentIDStore, result.id)
+    const doc = orama.data.docs.docs[originalID]
+
+    if (doc) {
+      if (!includeVectors) {
+        doc[property] = null
+      }
+
+      const newDoc: Result<ResultDocument> = {
+        id: result.id,
+        score: result.score,
+        document: doc
+      }
+      docs[i] = newDoc
+    }
+  }
+
+  const timeEnd = await getNanosecondsTime()
+  const elapsedTime = timeEnd - timeStart
+
+  return {
+    count: results.length,
+    hits: docs.filter(Boolean),
+    elapsed: {
+      raw: Number(elapsedTime),
+      formatted: await formatNanoseconds(elapsedTime)
+    }
+  }
+}
+
 
 async function fullTextSearch<T extends AnyOrama, ResultDocument = TypedDocument<T>>(orama: T, params: SearchParamsFullText<T, ResultDocument>, language?: string): Promise<Results<ResultDocument>> {
     const timeStart = await getNanosecondsTime()
